% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVT.R
\docType{data}
\name{CVT}
\alias{CVT}
\title{CVT Data}
\format{
A dataframe with 167 rows 6 variables
\describe{
\item{r1}{otolith length}
\item{otolith.width}{otolith width}
\item{otolith.weight}{otolith weight}
\item{fish.length}{fish lenght}
\item{fish.weight}{fish weight}
\item{age}{ age}
\item{sex}{ sex}
}
}
\usage{
data(CVT)
}
\description{
CVT Data
}
\examples{
 data("CVT")
}
\keyword{datasets}
