% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otolith.R
\docType{data}
\name{otolith}
\alias{otolith}
\title{Otolith Data}
\format{
A dataframe with 167 rows 6 variables
\describe{
\item{otolith.length}{otolith length}
\item{otolith.width}{otolith width}
\item{otolith.weight}{otolith weight}
\item{fish.length}{fish lenght}
\item{fish.weight}{fish weight}
\item{age}{ age}
\item{sex}{ sex}
}
}
\source{
{Doğan Y. Şen D.,  Otolith Biometry-Fish Lenth Relationship in Capoeta trutta Inhabiting Keban Dam Lake}
}
\usage{
data(otolith)
}
\description{
The data related to otolith bone of fishes was collected from Elazığ Keban Dam Lake (November 2011-December 2012),
which was a part of the data by given Doğan and Şen(2017). The data containing otolith length, otolith width, otolith weight,
fish length, fish width, age and sex.
}
\examples{
 data("otolith")
}
\keyword{datasets}
